Extract unit test
